Board Setup Operations!
This is the first round of Board class refactoring, complete with methods that SHOULD support creating any shape map (that is connected of course) around any centerpoint/local origin.
There's a bit of math that goes into orientating and defining the TileMap array of TileScripts, but it SHOULD work. I haven't tested yet, but I'm planning more lightweight refactorings and methods to add to the Board's utility.

:dromedary_camel: