Drafting the BattlePiece Abstraction
It's namespaced and everything! Here, I included a diagram to describe exactly what I wanted from my system. I have my BattleAvatar derivatives reading from their ScriptableObjects and all that, but I have yet to put some of the bells and whistles onto the BattleNPO derivatives.
That said, I have done work today that is almost equal to the work and learning I've done for the last few months. That's tremendous for me, and I'm really quite pleased with how things are coming along.
Next up, I want to focus on the Command Pattern and/or the UI.

:dromedary_camel: