CORRECTED MERGE MISTAKE
It turns out that the merge I had done had somehow moved all of the correct files to another cloned Assets file. After a bit of confusion and file management, I have created a stable new build of the project, within this repository, with all of the reviewed and polished code. Thank goodness I got that sorted out!

:dromedary_camel: