Implementing Initial Move + Elemental Things!
An important part of the game is elemental-focused BattleAvatars. Some are better against others, and vice versa, based on their Elemental type. Thus, I've implemented a rough implementation of these concepts.

As well, I'm working on the MoveData ScriptableObject, that is a container for all the moves that we have in the game. A move has two important characteristics: base damage and type. I've created the ScriptableObject to contain such data! I've gotten a bit of practice with the backing field and propery combo. I wonder if there's a hotkey for that, or a refactor option with that.

:dromedary_camel: