Tidying TileScript, Adding Methods
I'm just making tidy edits in Atom, so I don't know if they're super sound, but I'll keep doing it, it's good.
Chances are, I might not need to Raycast to find the GameObject on the field, instead using Vector2Int comparison, but that might not be as fast. I guess I have either method to do.

:dromedary_camel: