Trigger Collider Idea!
If I have a trigger collider that can check for GameObjects above the tile, I won't need to have to implement a crazy system that RayCasts to check it!
That said, I have yet to implement the code, but I have made sure to make the collider smaller than the Tile so I ONLY get GameObjects on the Tile.

:dromedary_camel: