Added Kinematic Rigidbodies for Tiles
As the title says, I added some Rigidbodies on all the Tiles that help the Trigger Colliders detect GameObjects on.
The only problem is that as my logging  notes, a GameObject between two colliders is technically within both colliders. However, the collider that has most recently been entered is set as the Tile's new ObjectOn, which is technically current. Finally, this "bug" has no impact on the way my game will run, since it is turn-based and all GO's on the board are locked onto their Tiles when any sort of checking occurs.
As well, I added a line for when I rethink my BattleGuys hierarchy (characters and Cards), to add a TileOn field in the BattleGuy and set it alongside our TIles' colliders.

:dromedary_camel: