Defining Adjacents
This was a difficult and fun problem that kept me up last night. I was wrestling NullReferenceExceptions and refactoring, and I ended up just checking whether or not a neighboring TileController exists instead of trying to store a live reference to it (that turns out to be illegal or extremely volatile, referencing scene objects).

:dromedary_camel: