Deleting old BP Derivatives
The old BattlePiece Derivatives were quite strange with their implications within my old system. Out with the old, in with the new, so I'll have to rethink how these fellas work as well-
IF,
I am to keep them around at all. BattleController is growing to be quite strong on its own, having ALL of the capabilities to Move, Attack, and do more, but restricts itself based on the type of its SOData. :]
I'll have to see what I can't do about relegating certain functionalities to SOData, in the case of Interactables and things like that, should I choose to implement them.

:dromedary_camel: