Updating BoardPosition, Hiding fields in Editor
Firstly, BattleControllers now have a method to have their BoardPosition updated, and set to the array location to the TIleController they're on.

Secondly, I've begun to consider hiding unnecessary information that is serialized in the editor. I've started this with BattleController, with fields that are updated internally and don't need setting in-editor.

:dromedary_camel: