[cosmetic}] Codemaid AND a new Attribute
As well as restricting what is necessary to be seen in the editor, I've begun to realize how volatile having multiple component scripts upon one object can be. I've begun implementing this attribute, [DisallowMultipleComponent], to avoid any monkey business that might happen if I or anyone else is careless with their design in the editor.

:dromedary_camel: