Implemented Tile Checking & Coroutine Safety within my Move() method!
I'm so happy that I can solve these problems on my own! :)

So the first issue I foresaw was using Move() and going somehow in an area that was out of the array, or to a Tile that didn't exist within the bounds of the TileMap array.

I also wrestled with some issues about the Lerp CRT with world space vs Board Position. I solved this with my BoardManager reference, however.

I also noticed when testing that I could  not start the same CRT when moving at the same time. I implemented a boolean called m_IsMoving to ONLY move when the CRT is finished. This can also mirror a boolean value in the Animator when I have models or sprites to work with, in terms of doing a movement sprite. :]

:dromedary_camel: