Implementing CardController Attack and HP methods!
These are obviously subject to change if necessary, but these methods outline what exactly I'm trying to accomplish through abstraction of the BattleController superclass. For example, I can CardController.Attack() any BattleController, because it will have to implement (or throw and exception or something) some way to be attacked, if possible. The beauty of abstraction and inheritance!

:dromedary_camel: